Move charge link command objects to new location in infrastructure
