+++ source/GreenEnergyHub.Charges/source/GreenEnergyHub.Charges.Infrastructure/ChargeLinks/Cim/ChargeLinkCommandConverter.cs
@@ -18,13 +18,13 @@ using System.Threading.Tasks;
 using System.Xml;
 using GreenEnergyHub.Charges.Domain.ChargeLinkCommands;
 using GreenEnergyHub.Charges.Domain.MarketParticipants;
-using GreenEnergyHub.Charges.Infrastructure.ChargeLinks.Cim;
 using GreenEnergyHub.Charges.Infrastructure.Correlation;
 using GreenEnergyHub.Charges.Infrastructure.MarketDocument.Cim;
+using GreenEnergyHub.Charges.Infrastructure.Messaging.Serialization;
 using GreenEnergyHub.Messaging.Transport;
 using NodaTime;
 
-namespace GreenEnergyHub.Charges.Infrastructure.Messaging.Serialization.Commands
+namespace GreenEnergyHub.Charges.Infrastructure.ChargeLinks.Cim
 {
     public class ChargeLinkCommandConverter : DocumentConverter
     {
