+++ source/GreenEnergyHub.Charges/source/GreenEnergyHub.Charges.Infrastructure/ChargeLinks/Cim/ChargeLinkCommandDeserializer.cs
@@ -17,9 +17,10 @@ using System.Threading;
 using System.Threading.Tasks;
 using System.Xml;
 using GreenEnergyHub.Charges.Domain.ChargeLinkCommands;
+using GreenEnergyHub.Charges.Infrastructure.Messaging.Serialization;
 using GreenEnergyHub.Messaging.Transport;
 
-namespace GreenEnergyHub.Charges.Infrastructure.Messaging.Serialization.Commands
+namespace GreenEnergyHub.Charges.Infrastructure.ChargeLinks.Cim
 {
     public class ChargeLinkCommandDeserializer : MessageDeserializer<ChargeLinkCommand>
     {
