+++ source/GreenEnergyHub.Charges/source/GreenEnergyHub.Charges.Infrastructure/ChargeLinks/Cim/CimChargeLinkCommandConstants.cs
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-namespace GreenEnergyHub.Charges.Infrastructure.Messaging.Serialization.Commands
+namespace GreenEnergyHub.Charges.Infrastructure.ChargeLinks.Cim
 {
     /// <summary>
     /// Strings used in CIM/XML for elements, namespaces or attributes that we need to
